moved FeroxMessage from lib.rs
